requirements.txt
django
djangorestframework
djangorestframework-simple-jwt